% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumsim.R
\name{eval_param_rsimsum}
\alias{eval_param_rsimsum}
\title{Vector of evaluated parameters in the simulation study that will be summarised
using the rsimsum package}
\usage{
eval_param_rsimsum()
}
\value{
vector of length 12 with the names of the to be evaluated parameters
and the name of their monte carlo standerd error
}
\description{
Vector of evaluated parameters in the simulation study that will be summarised
using the rsimsum package
}
