% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumsim.R
\name{fill_stat_and_mcse}
\alias{fill_stat_and_mcse}
\title{Fill summary with the evaluated parameter 'stat' and its mcse for a specific
scen_no and method using the simsum object from the package rsimsum}
\usage{
fill_stat_and_mcse(summary, simsum, scen_no, stat, method)
}
\arguments{
\item{summary}{the summary object initialised by init_summary()}

\item{simsum}{a simsum object created by the package rsimsum for a specific
scenario no.}

\item{scen_no}{scenario no. of the scenario that's evaluated}

\item{stat}{one of the parameters in eval_param_rsimsum()}

\item{method}{the method that's evaluated, either "uncor" / "mecor" / "simex"}
}
\value{
this function will fill out the specific cells with 'stat' and its
'stat'_mcse in the summary object (for a given scen_no and method)
}
\description{
Fill summary with the evaluated parameter 'stat' and its mcse for a specific
scen_no and method using the simsum object from the package rsimsum
}
