% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{generate_data}
\alias{generate_data}
\title{Generate data for simulation study}
\usage{
generate_data(seed_no, nobs, gamma, omega_sq, tau_sq, sigma_sq)
}
\arguments{
\item{seed_no}{No. of seed}

\item{nobs}{Number of observations}

\item{gamma}{Cov(X,Z) / Var(X)}

\item{omega_sq}{Var(X|Z)}

\item{tau_sq}{Extra variance due to measurement error in error-prone measure X_star}

\item{sigma_sq}{Var(Y|X,Z)}
}
\value{
Data.frame with variables Z, X, X_star_1, X_star_2, X_star_3 and Y
}
\description{
Generate data for simulation study
}
