% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm_logreg.R
\name{generate_data_logreg}
\alias{generate_data_logreg}
\title{Generate data for simulation study PART II (log reg)}
\usage{
generate_data_logreg(seed_no, nobs, gamma, tau_sq, nrep = 2, phi, beta = 0.1)
}
\arguments{
\item{seed_no}{seed no. used to generate data}

\item{nobs}{number of observations}

\item{gamma}{effect of Z on X}

\item{tau_sq}{extra variance due to measurement error in error-prone measure X_star}

\item{nrep}{number of replicates of X_star (defaults to 2)}

\item{phi}{effect of Z on Y (given X)}

\item{beta}{estimand (defaults to 0.05)}
}
\value{
Data.frame with variables Z, X, X_star_1, ..., X_star_nrep and Y
}
\description{
Generate data for simulation study PART II (log reg)
}
