% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm_logreg.R
\name{generate_data_logreg_scen_no}
\alias{generate_data_logreg_scen_no}
\title{Generate data part II (logreg) of a specific scenario number specified in input data available
in the package}
\usage{
generate_data_logreg_scen_no(seed_no, scen_no)
}
\arguments{
\item{seed_no}{seed no. used to generate data}

\item{scen_no}{scenario no. in the input to generate data from}
}
\value{
a data.frame with the data containing the variables Z1, Z2, X, X_star_1,
..., X_star_nrep and Y
}
\description{
Generate data part II (logreg) of a specific scenario number specified in input data available
in the package
}
