% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perrier.R
\name{generate_data_perrier}
\alias{generate_data_perrier}
\title{Generate data equal to the simulation study conducted by Perrier et al.}
\usage{
generate_data_perrier(seed_no, nobs = 3000, nrep = 2, tau_sq = 2/3)
}
\arguments{
\item{seed_no}{seed no. used to generate data}

\item{nobs}{number of observations (sample size)}

\item{nrep}{number of replicates of X_star (defaults to 2)}

\item{tau_sq}{extra variance due to measurement error in error-prone measure X_star (defaults to 2/3)}
}
\description{
Generate data equal to the simulation study conducted by Perrier et al.
}
