% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{generate_data_scen_no}
\alias{generate_data_scen_no}
\title{Generate data of a specific scenario number specified in input data available
in the package}
\usage{
generate_data_scen_no(seed_no, scen_no)
}
\arguments{
\item{seed_no}{seed no. used to generate data}

\item{scen_no}{scenario no. in the input to generate data from}
}
\value{
a data.frame with the data containing the variables Z, X, X_star_1,
X_star_2, X_star_3 and Y
}
\description{
Generate data of a specific scenario number specified in input data available
in the package
}
