% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_logreg.R
\name{get_crude_logreg}
\alias{get_crude_logreg}
\title{Get effect of crude model (Y ~ X)}
\usage{
get_crude_logreg(data)
}
\arguments{
\item{data}{data used to estimate crude model}
}
\value{
named vector with crude_effect
}
\description{
Get effect of crude model (Y ~ X)
}
