% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_r_squared}
\alias{get_r_squared}
\title{Get R-squared of the outcome model (Y ~ X + Z)}
\usage{
get_r_squared(data)
}
\arguments{
\item{data}{data used to estimate R-squared}
}
\value{
named vector with r_squared, the r squared of the outcome model
}
\description{
Get R-squared of the outcome model (Y ~ X + Z)
}
