% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_logreg.R
\name{get_r_squared_logreg}
\alias{get_r_squared_logreg}
\title{Get R-squared of the outcome model (Y ~ X + Z1)}
\usage{
get_r_squared_logreg(data)
}
\arguments{
\item{data}{data used to estimate R-squared}
}
\value{
named vector with r_squared, the r squared of the outcome model
}
\description{
Get R-squared of the outcome model (Y ~ X + Z1)
}
