% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perrier.R
\name{get_r_squared_perrier}
\alias{get_r_squared_perrier}
\title{#' Get R-squared of the outcome model (Y ~ X)}
\usage{
get_r_squared_perrier(data)
}
\arguments{
\item{data}{data used to estimate R-squared}
}
\value{
named vector with r_squared, the r squared of the outcome model
}
\description{
#' Get R-squared of the outcome model (Y ~ X)
}
