% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumsim.R
\name{init_summary}
\alias{init_summary}
\title{Initialise summary object}
\usage{
init_summary(scen_nos, input)
}
\arguments{
\item{scen_nos}{scen no.'s to be used in the summary object}

\item{input}{input of simulation study}
}
\value{
a data.frame with length(scen_nos) rows and length(eval_param_rsimsum())
+1 columns. See function eval_param_rsimsum() for evaluated parameters in
summary. Additionally, a column for the estimated r-squared is added.
}
\description{
Initialise summary object
}
