% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{input}
\alias{input}
\title{Input of Simulation Study}
\format{
A data frame with 23 rows and 10 variables:
\describe{
  \item{scen_no}{scenario no.}
  \item{nobs}{number of observations (sample size)}
  \item{tau_sq}{\eqn{\tau^2}}
  \item{beta}{estimand}
  \item{reliability}{Var(active_energy_expenditure) / Var(active_energy_expenditure_star)}
}
}
\usage{
input
}
\description{
A dataset containing the 48 scenarios that are used to input the simulation
study. The data generating mechanisms used in this study are:\cr
active_energy_expenditure ~ N(1, 0.25) \cr
active_energy_expenditure_star|active_energy_expenditure ~ N(active_energy_expenditure, \eqn{\tau^2}) \cr \cr
perc_lean_bm ~ N(80 + beta * active_energy_expenditure, 5)
}
\examples{
data("input", package = "woorc")
}
\keyword{datasets}
