% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{input_biasvsvar}
\alias{input_biasvsvar}
\title{Input of Simulation Study // Bias vs Variance Tradeoff RegCal}
\format{
A data frame with 30 rows and 6 variables:
\describe{
  \item{scen_no}{scenario no.}
  \item{nobs}{number of observations (sample size)}
  \item{tau_sq}{\eqn{\tau^2}}
  \item{beta}{estimand}
  \item{sigma_sq}{\eqn{\sigma^2}}
  \item{reliability}{Var(active_energy_expenditure) / Var(active_energy_expenditure_star)}
}
}
\usage{
input_biasvsvar
}
\description{
A dataset containing the 30 scenarios that are used to input the simulation
study illustrating the bias vs variance tradeoff between regression calibration
and the ordinary least squares estimator not corrected for measurement error.
The data generating mechanisms used in this study are:\cr
active_energy_expenditure ~ N(1, 0.25) \cr
active_energy_expenditure_star|active_energy_expenditure ~ N(active_energy_expenditure, \eqn{\tau^2}) \cr \cr
perc_lean_bm ~ N(80 + beta * active_energy_expenditure, \eqn{\sigma^2})
}
\examples{
data("input_biasvsvar", package = "woorc")
}
\keyword{datasets}
