% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{input_lowrel}
\alias{input_lowrel}
\title{Input of Simulation Study // Performance RegCal when Reliability is Low}
\format{
A data frame with 65 rows and 6 variables:
\describe{
  \item{scen_no}{scenario no.}
  \item{nobs}{number of observations (sample size)}
  \item{tau_sq}{\eqn{\tau^2}}
  \item{beta}{estimand}
  \item{sigma_sq}{\eqn{\sigma^2}}
  \item{reliability}{Var(active_energy_expenditure) / Var(active_energy_expenditure_star)}
}
}
\usage{
input_lowrel
}
\description{
A dataset containing the 65 scenarios that are used to input the simulation
study illustrating the performance of regression calibration when reliability of the
error prone exposure is low. The data generating mechanisms used in this study are:\cr
active_energy_expenditure ~ N(1, 0.25) \cr
active_energy_expenditure_star|active_energy_expenditure ~ N(active_energy_expenditure, \eqn{\tau^2}) \cr \cr
perc_lean_bm ~ N(80 + beta * active_energy_expenditure, \eqn{\sigma^2})
}
\examples{
data("input_lowrel", package = "woorc")
}
\keyword{datasets}
