% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{input_perrier}
\alias{input_perrier}
\title{Input of Simulation Study Conducted by Perrier et al.}
\format{
A data frame with 9 rows and 4 variables:
\describe{
  \item{scen_no}{scenario no.}
  \item{nobs}{number of observations (sample size)}
  \item{nrep}{number of replicates}
  \item{tau_sq}{\eqn{\tau^2}}
  \item{beta}{estimand (equal to 100)}
}
}
\usage{
input_perrier
}
\description{
A dataset containing 9 scenarios that are used to input the simulation
study conducted by Perrier et al. The data generating mechanisms used in this study are:\cr
X ~ N(0, 1) \cr
X_star|X ~ N(X, \eqn{\tau^2}) \cr
Y ~ N(14900 - 100 X, \eqn{1650^2}) \cr
}
\examples{
data("input_perrier", package = "simexvsmecor")
}
\keyword{datasets}
