% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procoutput.R
\name{make_output_long}
\alias{make_output_long}
\title{Reshape the output of the simulation study to long format}
\usage{
make_output_long(scen_no, output_dir = "./output/")
}
\arguments{
\item{scen_no}{scenario no.}

\item{output_dir}{directory where raw output files are to be found, defaults
to "./output/". Here, files named 'scen_no#' should be located}
}
\value{
a data.frame with 6 columns: 'r_squared', 'method', 'effect', 'se',
'ci.lower' and 'ci.upper'. The number of rows is 3 times the number of
replications used in the simulation study (number of replications is equal to
the number of rows in the raw output)
}
\description{
Reshape the output of the simulation study to long format
}
