% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{perform_mecor}
\alias{perform_mecor}
\title{Perform measurement error correction by means of mecor}
\usage{
perform_mecor(data)
}
\arguments{
\item{data}{data to be used}
}
\value{
numeric vector with the coefficient of the corrected analysis, the
standard error of the estimated coefficient and the confidence interval based
on the bootstrap
}
\description{
Perform measurement error correction by means of mecor
}
