% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_logreg.R
\name{perform_simex_logreg}
\alias{perform_simex_logreg}
\title{Perform measurement error correction by means of simex}
\usage{
perform_simex_logreg(data)
}
\arguments{
\item{data}{data to be used}
}
\value{
numeric vector with the coefficient of the corrected analysis, the
standard error of the estimated coefficient and the confidence interval based
on the jackknife variance component in simex
}
\description{
Perform measurement error correction by means of simex
}
