% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_logreg.R
\name{perform_uncor_logreg}
\alias{perform_uncor_logreg}
\title{Perform the uncorrected analysis PART II (logreg)}
\usage{
perform_uncor_logreg(data)
}
\arguments{
\item{data}{data to be used}
}
\value{
numeric vector with the coefficient of the uncorrected analysis, the
standard error of the estimated coefficient and the confidence interval
}
\description{
Perform the uncorrected analysis PART II (logreg)
}
