% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perrier.R
\name{perform_uncor_perrier}
\alias{perform_uncor_perrier}
\title{Perform the uncorrected analysis}
\usage{
perform_uncor_perrier(data)
}
\arguments{
\item{data}{data to be used}
}
\value{
numeric vector with the coefficient of the uncorrected analysis, the
standard error of the estimated coefficient and the confidence interval
}
\description{
Perform the uncorrected analysis
}
