% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procoutput.R
\name{process_output}
\alias{process_output}
\title{Tidy output of simulation study}
\usage{
process_output(
  scen_nos,
  output_dir = "./output/",
  processed_dir = "./output/processed/"
)
}
\arguments{
\item{scen_nos}{scen no.'s that should be tidied up/ processed}

\item{output_dir}{directory where raw output files are to be found, defaults
to "./output/". Here, files named 'scen_no#' should be located}

\item{processed_dir}{directory where the processed output should be saved,
defaults to "./output/processed/"}
}
\value{
for each scen no. a .rds file named 'scen_no#.rds" is created in
processed_dir with the tidy output of the sim study in long format
}
\description{
Tidy output of simulation study
}
