% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_logreg.R
\name{regcal_logreg}
\alias{regcal_logreg}
\title{Perform measurement error correction by means of regcal for logreg}
\usage{
regcal_logreg(data, var)
}
\arguments{
\item{data}{data to be used}

\item{var}{assumed variance of measurement error}
}
\value{
numeric vector with the coefficient of the corrected analysis
}
\description{
Perform measurement error correction by means of regcal for logreg
}
