% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procoutput.R
\name{reshape_output}
\alias{reshape_output}
\title{Reshape one evaluated parameter of the simulation output to long format}
\usage{
reshape_output(output_one_par, v.names)
}
\arguments{
\item{output_one_par}{data.frame with one row 'id' and 3 columns with an
evaluated parameter preceding "uncor."/ "mecor."/ "simex.", optionally this
data.frame may include other columns}

\item{v.names}{name of the evaluated parameter, can be either "effect"/ "se"/
"ci.lower"/ ci.upper"}
}
\value{
a data.frame with at least 4 columns: 'id', 'method' indicating
either "uncor"/ "mecor"/ "simex", and the evaluated parameters equal to
'v.names'
}
\description{
Reshape one evaluated parameter of the simulation output to long format
}
