% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runsim.R
\name{run_sim}
\alias{run_sim}
\title{Run simulation study}
\usage{
run_sim(nrep, scen_no, output_file_loc = "./")
}
\arguments{
\item{nrep}{number of replications}

\item{scen_no}{scenarios numbers to be used}

\item{output_file_loc}{directory where output of simulation will be saved}
}
\value{
a file named scen_no_* with the output will be saved in
output_file_loc, which constitutes of the uncorrected effect, the corrected
effect by means of mecor and the corrected effect by means of simex and the
seed no.
}
\description{
Run simulation study
}
