% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{save_output}
\alias{save_output}
\title{Save output of one single replication of simulation study}
\usage{
save_output(output, file)
}
\arguments{
\item{output}{the output to be saved}

\item{file}{the file where the output needs to be saved}
}
\description{
Save output of one single replication of simulation study
}
