% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumsim.R
\name{summarise_one_scen_no}
\alias{summarise_one_scen_no}
\title{Summarise one scenario no.}
\usage{
summarise_one_scen_no(summary, scen_no, processed_dir = "./output/processed/")
}
\arguments{
\item{summary}{data.frame initiated by init_summary() to be filled with
summary parameters}

\item{scen_no}{scen no. to be summarised}

\item{processed_dir}{directory where processed simulation output is to be
found, defaults to "./output/processed/"}
}
\value{
this function will fill out the rows in the summary object of the
corresponding scen_no
}
\description{
Summarise one scenario no.
}
