% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumsim.R
\name{summarise_sim}
\alias{summarise_sim}
\title{Summarise simulation study}
\usage{
summarise_sim(scen_nos, use_input, processed_dir = "./output/processed/")
}
\arguments{
\item{scen_nos}{scen no.'s to be summarised}

\item{use_input}{input of simulation study}

\item{processed_dir}{directory where the processed files of the simulation is to be found
found}
}
\value{
a filled data.frame that is initiated using the function
init_summary(), and additionally a column with the percentage bias in each
scenario for each method
}
\description{
Summarise simulation study
}
