% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{generate_data}
\alias{generate_data}
\title{Generate data for simulation study}
\usage{
generate_data(
  seed_no,
  nobs,
  gamma,
  omega_sq,
  nrep = 3,
  tau_sq,
  sigma_sq,
  beta = 0.2
)
}
\arguments{
\item{seed_no}{seed no. used to generate data}

\item{nobs}{number of observations}

\item{gamma}{Cov(X,Z) / Var(X)}

\item{omega_sq}{Var(X|Z)}

\item{nrep}{number of replicates of X_star (defaults to 3)}

\item{tau_sq}{extra variance due to measurement error in error-prone measure X_star}

\item{sigma_sq}{Var(Y|X,Z)}

\item{beta}{estimand, defaults to 0.2}
}
\value{
Data.frame with variables Z, X, X_star_1, ..., X_star_nrep and Y
}
\description{
Generate data for simulation study
}
