% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumsim.R
\name{summarise_sim}
\alias{summarise_sim}
\title{Summarise simulation study}
\usage{
summarise_sim(scen_nos, output_dir = "./output/")
}
\arguments{
\item{scen_nos}{scen no.'s to be summarised}

\item{output_dir}{directory where the output of the simulation study is to be
found}
}
\value{
a filled data.frame that is initiated using the function
init_summary(), and additionally a column with the percentage bias in each
scenario for each method
}
\description{
Summarise simulation study
}
